Cosmetics.

+++ src/main/java/interfaces/OutputSubscriber.java
@@ -1,5 +1,5 @@
 package interfaces;
 
 public interface OutputSubscriber {
-    void setSelectionImage(int x, int y);
+    void setSelectionImage(double x, double y);
 }
